Automatically backed up by Learn

+++ app/controllers/application_controller.rb
@@ -61,7 +61,7 @@ class ApplicationController < Sinatra::Base
   post "/withdraw" do
     binding.pry
     @user = User.find(session[:user_id])
-    if @user.balance > params[:withdrawal].to_i
+    if @user.balance > params[:withdrawal].to_f
       @user.balance -= params[:withdrawal]
       redirect "/account"
     else
